backend/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework>=3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn
psycopg2-binary
pillow
django-extensions>=4.1
